IT_school_Rostelecom/Rostelekom3_5_bot_xls_exell/haystack-rag-app-main/requirements.txt
haystack-ai
datasets>=2.6.1
sentence-transformers>=4.1.0
transformers[torch,sentencepiece]
python-dotenv
huggingface_hub[hf_xet]
ollama-haystack
python-telegram-bot
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
psycopg2-binary>=2.9.0